revert dialog component changes... move onclose into individual dialog components
