+++ lib/dialog/index.tsx
@@ -1,25 +1,15 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-import * as S from '../state';
-
-import { closeDialog } from '../state/ui/actions';
-
-type DispatchProps = {
-  closeDialog: () => any;
-};
-
-type Props = DispatchProps;
-
-export class Dialog extends Component<Props> {
+export class Dialog extends Component {
   static propTypes = {
     children: PropTypes.node.isRequired,
     className: PropTypes.string,
     closeBtnLabel: PropTypes.string,
     hideTitleBar: PropTypes.bool,
     title: PropTypes.string,
+    onDone: PropTypes.func,
   };
 
   render() {
@@ -29,6 +19,7 @@ export class Dialog extends Component<Props> {
       hideTitleBar,
       title,
       children,
+      onDone,
     } = this.props;
 
     return (
@@ -43,12 +34,12 @@ export class Dialog extends Component<Props> {
             <div className="dialog-title-side" />
             <h2 className="dialog-title-text">{title}</h2>
             <div className="dialog-title-side">
-              {!!this.props.closeDialog && (
+              {!!onDone && (
                 <button
                   type="button"
                   aria-label="Close dialog"
                   className="button button-borderless"
-                  onClick={this.props.closeDialog}
+                  onClick={onDone}
                 >
                   {closeBtnLabel}
                 </button>
@@ -63,10 +54,4 @@ export class Dialog extends Component<Props> {
   }
 }
 
-const mapDispatchToProps: S.MapDispatch<DispatchProps> = dispatch => ({
-  closeDialog: () => {
-    dispatch(closeDialog());
-  },
-});
-
-export default connect(null, mapDispatchToProps)(Dialog);
+export default Dialog;
